requirements.txt
# Dependencies
# Core pipeline
dlt
dbt-core
dbt-postgres
requests>=2.28.0
psycopg2-binary

# Data science
pandas>=1.3.0
numpy>=1.21.0

# Geolocation 
geopy>=2.2.0

# Environment 
python-dotenv>=0.19.0

# Additional 
pathlib>=1.0.1